+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ---
 
 ## [0.2.0]
+### Added
+* [#2]: validation of passed modifiers.
+
 ### Changed
 * [#1]: **BREAKING CHANGE**: removed the `tslib` dependency.
 
@@ -15,5 +18,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Yay, the first published version!
 
 [#1]: https://github.com/Comandeer/kreda/issues/1
+[#2]: https://github.com/Comandeer/kreda/issues/2
 
 [0.2.0]: https://github.com/Comandeer/kreda/compare/v0.1.0...v0.2.0